requirements.txt
langchain>=0.1.0
langchain-ollama>=0.1.0
langchain-community>=0.0.20
docx2txt
pypdf
fastapi
uvicorn
python-multipart
beautifulsoup4
PyYAML
langchain-openai